+++ tests/testthat/test-count-tally.r
@@ -28,15 +28,6 @@ test_that("name must be string", {
   expect_error(count(df, x, name = letters), "string")
 })
 
-test_that("weights by n if present, unless overridden", {
-  df <- tibble(g = c(1, 2), n = c(1, 3))
-  expect_message(out <- count(df, g), "weighting variable")
-  expect_equal(out, tibble(g = c(1, 2), n = c(1, 3)))
-
-  expect_message(out <- count(df, g, wt = n()), NA)
-  expect_equal(out, tibble(g = c(1, 2), n = c(1, 1)))
-})
-
 test_that("output includes empty levels with .drop = FALSE", {
   df <- tibble(f = factor("b", levels = c("a", "b", "c")))
   out <- count(df, f, .drop = FALSE)
@@ -78,14 +69,30 @@ test_that("works with dbplyr", {
   expect_equal(df2, tibble(x = c(1, 1, 1, 2, 2), n = c(3, 3, 3, 2, 2)))
 })
 
-test_that("can chain together multiple counts", {
+test_that("can only explicitly chain together multiple tallies", {
   verify_output(test_path("test-count-tally.txt"), {
     df <- data.frame(g = c(1, 1, 2, 2), n = 1:4)
-    df %>% count(g)
-    df %>% count(g, wt = n) %>% count()
+    df %>% count(g, wt = n)
+    df %>% count(g, wt = n) %>% count(wt = n)
 
     df %>% count(n)
   })
+})
+
+test_that("tally() automatically weights; count() does not", {
+  df <- tibble(g = c(1, 2), n = c(1, 3))
+  out <- count(df, g)
+  expect_equal(out, tibble(g = c(1, 2), n = c(1, 1)))
+
+  expect_message(out <- tally(group_by(df, g)), "weighting variable")
+  expect_equal(out, tibble(g = c(1, 2), n = c(1, 3)))
+
+  # But can override the defaults
+  expect_message(out <- count(df, g, wt = guess_wt()), "weighting variable")
+  expect_equal(out, tibble(g = c(1, 2), n = c(1, 3)))
+
+  out <- tally(group_by(df, g), wt = NULL)
+  expect_equal(out, tibble(g = c(1, 2), n = c(1, 1)))
 })
 
 # tally -------------------------------------------------------------------