+++ NEWS.md
@@ -1,6 +1,9 @@
 # dplyr (development version)
 
-* `count()` now displays the correct the hint to force a count, `wt = n()` 
+* `count()` no longer automatically weights by column `n` if present.
+  `count()` and `tally()` now have clearly different defaults for `wt`.
+
+* `count()` now displays the correct the hint to force a count, `wt = NULL` 
   (#5324).
 
 * `coalesce()` now supports data frames correctly (#5326).